+++ packages/grid_cockpit_ui/test/source/tree_source_test.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:flutter_test/flutter_test.dart';
 import 'package:grid_cockpit_ui/grid_cockpit_ui.dart';
-import 'package:grid_diagnostics_contract/grid_diagnostics_contract.dart';
+import 'package:genesis_foundation/genesis_foundation.dart';
 
 import '../fixtures.dart';