+++ packages/grid_devtools/lib/src/projection/projection_tabs.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:flutter/material.dart';
 import 'package:grid_cockpit_ui/grid_cockpit_ui.dart';
-import 'package:grid_diagnostics_contract/grid_diagnostics_contract.dart';
+import 'package:genesis_foundation/genesis_foundation.dart';
 
 import '../events/events_panel.dart';
 import '../protocol/grid_exploration_client.dart';