+++ packages/grid_devtools/test/projection_tabs_test.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:grid_cockpit_ui/grid_cockpit_ui.dart';
 import 'package:grid_devtools/grid_devtools.dart';
-import 'package:grid_diagnostics_contract/grid_diagnostics_contract.dart';
+import 'package:genesis_foundation/genesis_foundation.dart';
 
 import 'fake_grid_exploration_client.dart';